requirements.txt
fastapi
uvicorn
httpx
asyncpg
sqlalchemy[asyncio]
python-dotenv
aiosqlite
apscheduler
python-dateutil
